feat(#220) - Fix passing of the connection to related models in case when accessing cached entities relations && merged traits into one.
